+++ repository/src/main/java/org/grnet/cat/repositories/registry/MotivationActorRepository.java
@@ -5,11 +5,12 @@ import jakarta.enterprise.context.ApplicationScoped;
 import org.grnet.cat.entities.Page;
 import org.grnet.cat.entities.PageQuery;
 import org.grnet.cat.entities.PageQueryImpl;
+import org.grnet.cat.entities.registry.MotivationActorId;
 import org.grnet.cat.entities.registry.MotivationActorJunction;
 import org.grnet.cat.repositories.Repository;
 
 @ApplicationScoped
-public class MotivationActorRepository  implements Repository<MotivationActorJunction, String> {
+public class MotivationActorRepository  implements Repository<MotivationActorJunction, MotivationActorId> {
 
     /**
      * Retrieves a page of Motivations.
@@ -32,10 +33,9 @@ public class MotivationActorRepository  implements Repository<MotivationActorJun
         return pageable;
     }
 
-    public boolean existsByMotivationAndActorAndVersion(String motivationId, String actorId,Integer lodMAV) {
-        return find(
-                "SELECT 1 FROM MotivationActorJunction m WHERE m.id.motivationId = ?1 AND m.id.actorId = ?2 AND m.id.lodMAV=?3",
-                motivationId, actorId,lodMAV
-        ).firstResultOptional().isPresent();
+    public boolean existsByMotivationAndActorAndVersion(String motivationId, String actorId, Integer lodMAV) {
+        return find("SELECT 1 FROM MotivationActorJunction m WHERE m.id.motivationId = ?1 AND m.id.actorId = ?2 AND m.id.lodMAV = ?3", motivationId, actorId, lodMAV)
+                .firstResultOptional()
+                .isPresent();
     }
 }